Spectator glitch fix
A player could get around the class limits/bans by joining spectator
then rejoining the team, and would be able to choose any class.

This fix checks a player's team and class every time they change class
or join a playing team (RED or BLU; spec/unassigned is ignored). If the
player is found to be playing a banned or duplicate class, the player is
forced onto a class that is both unbanned and unused. If such a class
cannot be found (if you're playing 7v7 for example), the player is moved
(back) to spectator.